Term frequency analysis
Using the TF/IDF formula I ran term frequency analysis on a set of job descriptions to discover which words are associated with the highest salaries. The final results are stored in a a table where each of the top 10 and bottom 10 words are paired with their correlation to salary.